requirements.txt
langchain-google-genai
fastapi
uvicorn
pymongo[srv]
python-dotenv
